Create pull request template
### What
Added markdown pull request template to the project. Followed these [GitHub instructions](https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/creating-a-pull-request-template-for-your-repository).

### Why
To ensure we capture the what/why of our pull requests. This will help provide context for changes to the codebase which can be really useful when reviewing code history.